+++ src/app.py
@@ -6,6 +6,7 @@ from src.api.controllers.core.checkup import checkup_handler
 from src.api.controllers.core.liveness import liveness_handler
 from src.api.controllers.core.readiness import readiness_handler
 from src.api.controllers.core.metrics import metrics_handler
+from src.api.controllers.embeddings import embeddings_handler
 from src.api.middlewares.app_context_middleware import AppContextMiddleware
 from src.api.middlewares.logger_middleware import LoggerMiddleware
 from src.configurations.configuration import get_configuration
@@ -13,13 +14,15 @@ from src.configurations.variables import get_variables
 from src.context import AppContext, AppContextParams
 from src.infrastracture.logger import get_logger
 from src.infrastracture.metrics.manager import MetricsManager
+from src.lib.vector_search_index_updater import VectorSearchIndexUpdater
 
 
 def create_app(context: AppContext) -> FastAPI:
     app = FastAPI(
         openapi_url="/documentation/json",
-        # docs_url=None,
-        redoc_url=None
+        redoc_url=None,
+        title="ai-rag-template",
+        version="0.3.1"
     )
 
     app.add_middleware(AppContextMiddleware, app_context=context)
@@ -31,6 +34,7 @@ def create_app(context: AppContext) -> FastAPI:
     app.include_router(metrics_handler.router)
 
     app.include_router(chat_completions_handler.router)
+    app.include_router(embeddings_handler.router)
 
     return app
 
@@ -52,6 +56,9 @@ if __name__ == '__main__':
     )
 
     application = create_app(app_context)
+
+    vector_search_index_updater = VectorSearchIndexUpdater(app_context)
+    vector_search_index_updater.update_vector_search_index()
 
     uvicorn.run(
         application,