+++ tests/src/application/assistance/chains/retriever_chain_test.py
@@ -68,7 +68,7 @@ def setup_test(
     return mock_similar_documents, inputs, chain
 
 @patch(
-    'langchain_mongodb.MongoDBAtlasVectorSearch._similarity_search_with_score',
+    'langchain_community.vectorstores.mongodb_atlas.MongoDBAtlasVectorSearch._similarity_search_with_score',
 )
 def test_call(
     similarity_search_with_score,
@@ -86,10 +86,11 @@ def test_call(
     assert result[chain.output_key] == [
         doc for doc, _ in mock_similar_documents]
     for doc, _ in mock_similar_documents:
+        # pylint: disable=E1136
         assert doc.metadata['score'] == 0.5
         
 @patch(
-    'langchain_mongodb.MongoDBAtlasVectorSearch._similarity_search_with_score',
+    'langchain_community.vectorstores.mongodb_atlas.MongoDBAtlasVectorSearch._similarity_search_with_score',
 )
 def test_call_with_max_distance(
     similarity_search_with_score,
@@ -108,7 +109,7 @@ def test_call_with_max_distance(
     assert post_filter_call_arg[0]["$match"]["score"]["$lte"] == 0.5
 
 @patch(
-    'langchain_mongodb.MongoDBAtlasVectorSearch._similarity_search_with_score',
+    'langchain_community.vectorstores.mongodb_atlas.MongoDBAtlasVectorSearch._similarity_search_with_score',
 )
 def test_call_with_min_distance(
     similarity_search_with_score,